# fix: improve comments performance