--- a/packages/components/src/DiscussionContainer/DiscussionContainer.stories.tsx
+++ b/packages/components/src/DiscussionContainer/DiscussionContainer.stories.tsx
@@ -31,6 +31,7 @@ export const Default: Story = {
         onMoreComments={() => null}
         onSubmit={() => null}
         onSubmitReply={() => Promise.resolve()}
+        isSubmitting={false}
         isLoggedIn={false}
       />
     )
@@ -51,6 +52,7 @@ export const NoComments: Story = {
         onMoreComments={() => null}
         onSubmit={() => null}
         onSubmitReply={() => Promise.resolve()}
+        isSubmitting={false}
         isLoggedIn={false}
       />
     )
@@ -73,6 +75,7 @@ export const LoggedIn: Story = {
         onMoreComments={() => null}
         onSubmit={() => null}
         onSubmitReply={() => Promise.resolve()}
+        isSubmitting={false}
         isLoggedIn={true}
       />
     )
@@ -95,6 +98,7 @@ export const Expandable: Story = {
         onMoreComments={() => null}
         onSubmit={() => null}
         onSubmitReply={() => Promise.resolve()}
+        isSubmitting={false}
         isLoggedIn={true}
       />
     )
@@ -121,6 +125,7 @@ export const WithReplies: Story = {
         onMoreComments={() => null}
         onSubmit={() => null}
         isLoggedIn={true}
+        isSubmitting={false}
         onSubmitReply={async (commentId, comment) =>
           alert(`reply to commentId: ${commentId} with comment: ${comment}`)
         }