## PR Checklist

- [x] - Commit [messages are descriptive](https://github.com/ONEARMY/community-platform/blob/master/CONTRIBUTING.md#--commit-style-guide), it will be used in our [Release Notes](https://github.com/ONEARMY/community-platform/releases/)
- [ ] - Unit and e2e tests for the changes that have been added (for bug fixes / features)

## PR Type

What kind of change does this PR introduce?

- [x] Bugfix (fixes an issue)
- [x] Feature (adds functionality)
- [x] Refactoring (no functional changes)

## What is the current behavior?

Adding comments is very slow (13 seconds in a particular case!).
Due to notifications, the more contributors it has, the worse it gets!

Lack of loading indicators makes it worse.

## What is the new behavior?

Does not await discussion metadata updates and user notifications.
Locally it reduces the time from 4s to 1s.
1 second should be constant now, so for prod it would reduce from 13s to 1s.

## Does this PR introduce a breaking change?
- [ ] Yes
- [x] No

## Git Issues

Closes #