@davehakkens 2024-08-06
Seems to work well, also nice loading indication